moved facebook configuration from itinerary file to facebook route
